+++ lib/informers/configs.rb
@@ -1,6 +1,6 @@
 module Informers
   class PretrainedConfig
-    attr_reader :model_type, :problem_type, :id2label
+    attr_reader :model_type, :problem_type, :id2label, :label2id
 
     def initialize(config_json)
       @is_encoder_decoder = false
@@ -8,6 +8,7 @@ module Informers
       @model_type = config_json["model_type"]
       @problem_type = config_json["problem_type"]
       @id2label = config_json["id2label"]
+      @label2id = config_json["label2id"]
     end
 
     def [](key)